app -> game refactor complete app.js
