+++ src/components/containers/GameContainer.js
@@ -32,11 +32,7 @@ class GameContainer extends React.Component {
           height: "100vh",
         }}
       >
-        <Gamescreen
-          map={map}
-          territories={this.props.history[0].territories}
-          players={players}
-        />
+        <Gamescreen map={map} territories={territories} players={players} />
         <Sidebar
           roundCounter={roundCounter}
           playerToGo={playerToGo}
